Unify ERROR_CODES imports and rename constant file
Consolidate ERROR_CODES imports from @insforge/shared-schemas across multiple backend modules by grouping it with other named imports. Rename packages/shared-schemas/src/error-contants.ts to error-constants.ts and update the package index export accordingly. Also: adjust VercelProvider to throw INTERNAL_ERROR for canceled uploads, remove a duplicate import in a unit test, and apply minor formatting/newline fixes in shared error-codes.
