+++ backend/src/api/routes/auth/custom-oauth.routes.ts
@@ -1,7 +1,6 @@
 import { Router, Request, Response, NextFunction } from 'express';
 import jwt from 'jsonwebtoken';
 import { AppError } from '@/api/middlewares/error.js';
-import { ERROR_CODES } from '@insforge/shared-schemas';
 import { successResponse } from '@/utils/response.js';
 import { AuthRequest, verifyAdmin } from '@/api/middlewares/auth.js';
 import logger from '@/utils/logger.js';
@@ -17,6 +16,7 @@ import {
   listCustomOAuthConfigsResponseSchema,
   oAuthInitRequestSchema,
   customOAuthKeySchema,
+  ERROR_CODES,
 } from '@insforge/shared-schemas';
 
 const router = Router();
