+++ backend/src/api/routes/compute/services.routes.ts
@@ -3,8 +3,7 @@ import { verifyAdmin, AuthRequest } from '@/api/middlewares/auth.js';
 import { ComputeServicesService } from '@/services/compute/services.service.js';
 import { successResponse } from '@/utils/response.js';
 import { AppError } from '@/api/middlewares/error.js';
-import { ERROR_CODES } from '@insforge/shared-schemas';
-import { createServiceSchema, updateServiceSchema } from '@insforge/shared-schemas';
+import { createServiceSchema, updateServiceSchema, ERROR_CODES } from '@insforge/shared-schemas';
 import { AuditService } from '@/services/logs/audit.service.js';
 import { SocketManager } from '@/infra/socket/socket.manager.js';
 import { DataUpdateResourceType, ServerEvents } from '@/types/socket.js';
