+++ backend/src/providers/email/smtp.provider.ts
@@ -1,11 +1,10 @@
 import nodemailer from 'nodemailer';
 import type Mail from 'nodemailer/lib/mailer';
 import { AppError } from '@/api/middlewares/error.js';
-import { ERROR_CODES } from '@insforge/shared-schemas';
 import { EmailTemplate } from '@/types/email.js';
 import { SmtpConfigService, RawSmtpConfig } from '@/services/email/smtp-config.service.js';
 import { EmailTemplateService } from '@/services/email/email-template.service.js';
-import { SendRawEmailRequest } from '@insforge/shared-schemas';
+import { SendRawEmailRequest, ERROR_CODES } from '@insforge/shared-schemas';
 import { EmailProvider } from './base.provider.js';
 import logger from '@/utils/logger.js';
