+++ backend/src/services/ai/embedding.service.ts
@@ -1,8 +1,7 @@
 import { OpenRouterProvider } from '@/providers/ai/openrouter.provider.js';
-import type { EmbeddingsRequest, EmbeddingsResponse } from '@insforge/shared-schemas';
+import { type EmbeddingsRequest, type EmbeddingsResponse, ERROR_CODES } from '@insforge/shared-schemas';
 import logger from '@/utils/logger.js';
 import { AppError } from '@/api/middlewares/error.js';
-import { ERROR_CODES } from '@insforge/shared-schemas';
 
 export class EmbeddingService {
   private static instance: EmbeddingService;
