+++ backend/src/services/auth/auth-config.service.ts
@@ -2,9 +2,8 @@ import { Pool } from 'pg';
 import picomatch from 'picomatch';
 import { DatabaseManager } from '@/infra/database/database.manager.js';
 import { AppError } from '@/api/middlewares/error.js';
-import { ERROR_CODES } from '@insforge/shared-schemas';
 import logger from '@/utils/logger.js';
-import type { AuthConfigSchema, UpdateAuthConfigRequest } from '@insforge/shared-schemas';
+import { type AuthConfigSchema, type UpdateAuthConfigRequest, ERROR_CODES } from '@insforge/shared-schemas';
 import { URL } from 'url';
 
 export class AuthConfigService {
