+++ backend/src/services/auth/oauth-config.service.ts
@@ -2,9 +2,8 @@ import { Pool } from 'pg';
 import { DatabaseManager } from '@/infra/database/database.manager.js';
 import { SecretService } from '@/services/secrets/secret.service.js';
 import { AppError } from '@/api/middlewares/error.js';
-import { ERROR_CODES } from '@insforge/shared-schemas';
 import logger from '@/utils/logger.js';
-import { OAuthConfigSchema, OAuthProvidersSchema } from '@insforge/shared-schemas';
+import { OAuthConfigSchema, OAuthProvidersSchema, ERROR_CODES } from '@insforge/shared-schemas';
 
 export interface CreateOAuthConfigInput {
   provider: OAuthProvidersSchema;
